+++ itda-front/src/components/AddProduct/AddProduct.tsx
@@ -1,11 +1,13 @@
 import Header from "components/common/Header";
 import AddProductForm from "./AddProductForm";
+import AddProductSubmitLayer from "./AddProductSubmitLayer";
 
 const AddProduct = () => {
   return (
     <>
       <Header />
       <AddProductForm />
+      <AddProductSubmitLayer />
     </>
   );
 };
