--- itda-front/src/components/AddProduct/AddProductService.tsx
@@ -1,5 +1,0 @@
-const AddProductService = () => {
-	return <div></div>;
-};
-
-export default AddProductService;
