+++ itda-front/src/components/common/TinyEditor.tsx
@@ -1,16 +1,16 @@
 import { Editor } from "@tinymce/tinymce-react";
 import { useRecoilState } from "recoil";
-import { addProductInfos } from "stores/AddProductAtoms";
+import { editorValue } from "stores/AddProductAtoms";
 
 export default function TinyEditor() {
-  const [value, setValue] = useRecoilState<any>(addProductInfos);
+  const [value, setValue] = useRecoilState<any>(editorValue);
 
   return (
     <Editor
       value={value.detailDescription}
       apiKey="0jy0itwdqb4xmmeblh148y5w6bd3j22tjmc7udno3ptkinxk"
       onEditorChange={(newValue, editor) => {
-        setValue({ ...value, detailDescription: newValue });
+        setValue({ ...value, newValue });
       }}
       init={{
         height: 500,
